test(input): migrate states
