+++ core/src/components/input/test/states/input.e2e.ts-snapshots/input-readonly-ios-ltr-Mobile-Chrome-linux.png
